+++ src/math.odin
@@ -2,10 +2,10 @@ package main
 
 import "core:math"
 
-Vec2 :: [2]f32
-Vec3 :: [3]f32
-Vec4 :: [4]f32
-Mat4 :: matrix[4, 4]f32
+Vec2 :: [2]f64
+Vec3 :: [3]f64
+Vec4 :: [4]f64
+Mat4 :: matrix[4, 4]f64
 
 IntRect :: struct {
     x: int,
